Unconditionally build experimental interfaces for use by hse lib. Move experimental interfaces into include/hse/experimental.h. Install experimental.h only if -Dexperimental=true.
