+++ cli/cli_util.h
@@ -21,14 +21,12 @@ kvdb_info_print(
     const char *const *  paramv,
     struct yaml_context *yc);
 
-#ifdef HSE_EXPERIMENTAL
 bool
 kvdb_storage_info_print(
     const char *         kvdb_home,
     const size_t         paramc,
     const char *const *  paramv,
     struct yaml_context *yc);
-#endif
 
 int
 kvdb_compact_request(const char *kvdb_home, const char *request_type, unsigned timeout_sec);
