+++ include/hse/hse.h
@@ -166,54 +166,6 @@ hse_fini(void);
 hse_err_t
 hse_kvdb_close(struct hse_kvdb *kvdb);
 
-#ifdef HSE_EXPERIMENTAL
-
-/** @brief Request a data compaction operation.
- *
- * In managing the data within an HSE KVDB, there are maintenance activities
- * that occur as background processing. The application may be aware that it is
- * advantageous to do enough maintenance now for the database to be as compact
- * as it ever would be in normal operation.
- *
- * See the function hse_kvdb_compact_status_get().
- *
- * <b>Flags:</b>
- * @arg HSE_KVDB_COMPACT_CANCEL - Cancel an ongoing compaction request.
- * @arg HSE_KVDB_COMPACT_SAMP_LWM - Compact to the space amp low watermark.
- *
- * @note This function is thread safe.
- *
- * @param kvdb: KVDB handle from hse_kvdb_open().
- * @param flags: Compaction flags.
- *
- * @remark @p kvdb must not be NULL.
- * @remark @p flags must not be 0.
- *
- * @returns Error status.
- */
-hse_err_t
-hse_kvdb_compact(struct hse_kvdb *kvdb, unsigned int flags);
-
-/** @brief Get status of an ongoing compaction activity.
- *
- * The caller can examine the fields of the hse_kvdb_compact_status struct to
- * determine the current state of maintenance compaction.
- *
- * @note This function is thread safe.
- *
- * @param kvdb: KVDB handle from hse_kvdb_open().
- * @param[out] status: Status of compaction request.
- *
- * @remark @p kvdb must not be NULL.
- * @remark @p status must not be NULL.
- *
- * @returns Error status.
- */
-hse_err_t
-hse_kvdb_compact_status_get(struct hse_kvdb *kvdb, struct hse_kvdb_compact_status *status);
-
-#endif
-
 /** @brief Create a KVDB.
  *
  * @note This function is not thread safe.
@@ -330,27 +282,6 @@ hse_kvdb_open(
  */
 hse_err_t
 hse_kvdb_storage_add(const char *kvdb_home, size_t paramc, const char *const *paramv);
-
-#ifdef HSE_EXPERIMENTAL
-
-/** @brief Get KVDB storage stats.
- *
- * Obtain the space usage statistics for a specified kvdb.
- *
- * @note This function is thread safe.
- *
- * @param kvdb: KVDB handle from hse_kvdb_open().
- * @param[out] info: KVDB storage config and stats.
- *
- * @remark @p kvdb must not be NULL.
- * @remark @p info must not be NULL.
- *
- * @returns Error status.
- */
-hse_err_t
-hse_kvdb_storage_info_get(struct hse_kvdb *kvdb, struct hse_kvdb_storage_info *info);
-
-#endif
 
 /** @brief Sync data in all of the referenced KVDB's KVSs to stable media.
  *
@@ -575,59 +506,6 @@ hse_kvs_prefix_delete(
     struct hse_kvdb_txn *txn,
     const void *         pfx,
     size_t               pfx_len);
-
-#ifdef HSE_EXPERIMENTAL
-
-/** @brief Probe for a prefix.
- *
- * Given a prefix, outputs how many matches were encountered - zero, one or
- * multiple.
- *
- * @note This function is thread safe.
- *
- * <b>Flags:</b>
- * @arg 0 - Reserved for future use.
- *
- * @param kvs: KVS handle from hse_kvdb_kvs_open().
- * @param flags: Flags for operation specialization.
- * @param txn: Transaction context (optional).
- * @param pfx: Prefix to be probed.
- * @param pfx_len: Length of @p pfx.
- * @param[out] found: Zero, one or multiple matches seen.
- * @param[in,out] keybuf: Buffer which will be populated with contents of first
- * seen key.
- * @param keybuf_sz: Size of @p keybuf.
- * @param[out] key_len: Length of first seen key.
- * @param[in,out] valbuf: Buffer which will be populated with value for @p
- * keybuf.
- * @param valbuf_sz: Size of @p valbuf.
- * @param[out] val_len: Length of the value seen.
- *
- * @remark @p kvs must not be NULL.
- * @remark @p pfx must not be NULL.
- * @remark @p pfx_len must be within the range of [1, HSE_KVS_KEY_LEN_MAX].
- * @remark @p found must not be NULL.
- * @remark @p keybuf_sz must be equal to HSE_KVS_KEY_LEN_MAX.
- * @remark @p val_len must not be NULL.
- *
- * @returns Error status.
- */
-hse_err_t
-hse_kvs_prefix_probe(
-    struct hse_kvs *            kvs,
-    unsigned int                flags,
-    struct hse_kvdb_txn        *txn,
-    const void *                pfx,
-    size_t                      pfx_len,
-    enum hse_kvs_pfx_probe_cnt *found,
-    void *                      keybuf,
-    size_t                      keybuf_sz,
-    size_t *                    key_len,
-    void *                      valbuf,
-    size_t                      valbuf_sz,
-    size_t *                    val_len);
-
-#endif
 
 /** @brief Put a key-value pair into KVS.
  *
@@ -1092,6 +970,13 @@ hse_kvs_cursor_update_view(struct hse_kvs_cursor *cursor, unsigned int flags);
 
 #ifdef __cplusplus
 }
+#endif
+
+/* [HSE_REVISIT] The following inclusion will be removed after the db_bench
+ * hse-python repos are fixed to include hse_experimental.h directly.
+ */
+#ifdef HSE_EXPERIMENTAL
+#include <hse/experimental.h>
 #endif
 
 #endif
