+++ lib/binding/meson.build
@@ -1,6 +1,5 @@
 binding_sources = files(
     'diag_kvdb_interface.c',
-    'kvdb_experimental_interface.c',
     'kvdb_interface.c',
     'kvdb_perfc.c',
 )
